Fix verified_by in physical verification migration

+++ src/database/migrations/20260624091617_create_pcr_physical_verifications.ts
@@ -10,7 +10,7 @@ export async function up(knex: Knex): Promise<void> {
       table.string('status', 50).notNullable().defaultTo('Pending');
       table.string('condition', 50).notNullable().defaultTo('OK');
       table.text('remarks').nullable();
-      table.bigInteger('verified_by').unsigned().nullable();
+      table.integer('verified_by').unsigned().nullable();
       table.timestamp('verified_at').nullable();
       table.timestamp('created_at').defaultTo(knex.fn.now());
       table.timestamp('updated_at').defaultTo(knex.fn.now());
@@ -20,7 +20,7 @@ export async function up(knex: Knex): Promise<void> {
 
       // Unique constraint for upsert
       table.unique(['ticket_id', 'part_code'], { indexName: 'uq_pcr_verification_ticket_part' });
-      
+
       // Index for search optimization
       table.index(['ticket_id'], 'idx_pcr_verification_ticket');
       table.index(['part_code'], 'idx_pcr_verification_part');
